# sstable: optimize seeks to use next