This is a WIP and has the changes for singleLevelIterator.SeekGE.
It requires strengthening the requirement that the user call SeekGE
with a key that respects the lower bound.

The SeqSeek benchmarks are better
```
name                                                 old time/op    new time/op    delta
LevelIterSeqSeekGEWithBounds/restart=16/count=5-16     661ns ± 1%     152ns ± 1%  -76.96%  (p=0.000 n=8+9)
MergingIterSeqSeekGEWithBounds/levelCount=5-16        1.97µs ± 2%    0.72µs ± 2%  -63.37%  (p=0.000 n=9+10)
```

Overall, between this and the preceding PR, the MergingIterSeqSeek
benchmark shows a 5x improvement.

I'd like some feedback before adding SeekLT, SeekPrefixGE, and the
twoLevelIterator changes.

The more complete microbenchmark results comparing with master show
some slowdown, possibly because of wasted work when there is poor
locality. I need to investigate some more.
```
name                                            old time/op    new time/op    delta
MergingIterSeekGE/restart=16/count=1-16            783ns ± 6%     771ns ± 2%     ~     (p=0.986 n=10+9)
MergingIterSeekGE/restart=16/count=2-16           1.58µs ± 3%    1.59µs ± 2%   +1.07%  (p=0.045 n=10+9)
MergingIterSeekGE/restart=16/count=3-16           2.36µs ± 1%    2.41µs ± 2%   +2.07%  (p=0.000 n=8+10)
MergingIterSeekGE/restart=16/count=4-16           3.21µs ± 3%    3.26µs ± 2%   +1.64%  (p=0.005 n=10+9)
MergingIterSeekGE/restart=16/count=5-16           4.07µs ± 3%    4.16µs ± 1%   +2.10%  (p=0.008 n=10+9)
MergingIterNext/restart=16/count=1-16             37.4ns ± 0%    36.8ns ± 0%   -1.59%  (p=0.000 n=9+8)
MergingIterNext/restart=16/count=2-16             60.1ns ± 2%    60.3ns ± 1%     ~     (p=0.195 n=10+10)
MergingIterNext/restart=16/count=3-16             76.5ns ± 1%    76.4ns ± 1%     ~     (p=0.984 n=10+9)
MergingIterNext/restart=16/count=4-16             88.6ns ± 1%    88.7ns ± 2%     ~     (p=0.826 n=10+9)
MergingIterNext/restart=16/count=5-16              104ns ± 1%     105ns ± 1%   +1.42%  (p=0.003 n=9+8)
MergingIterPrev/restart=16/count=1-16             50.5ns ± 2%    55.6ns ± 4%  +10.07%  (p=0.000 n=10+10)
MergingIterPrev/restart=16/count=2-16             74.2ns ± 1%    78.5ns ± 1%   +5.88%  (p=0.000 n=10+8)
MergingIterPrev/restart=16/count=3-16             90.9ns ± 2%    99.4ns ± 5%   +9.32%  (p=0.000 n=9+10)
MergingIterPrev/restart=16/count=4-16              101ns ± 0%     108ns ± 1%   +7.30%  (p=0.000 n=8+8)
MergingIterPrev/restart=16/count=5-16              117ns ± 3%     135ns ± 7%  +15.04%  (p=0.000 n=10+10)
MergingIterSeqSeekGEWithBounds/levelCount=5-16    3.68µs ±11%    0.75µs ± 3%  -79.58%  (p=0.000 n=10+9)

name                                                old time/op    new time/op    delta
LevelIterSeekGE/restart=16/count=5-16                 1.29µs ± 3%    1.48µs ±13%  +14.78%  (p=0.000 n=10+9)
LevelIterSeqSeekGEWithBounds/restart=16/count=5-16     892ns ± 2%     155ns ± 1%  -82.59%  (p=0.000 n=10+8)
LevelIterNext/restart=16/count=5-16                   19.0ns ± 2%    20.6ns ±11%   +8.64%  (p=0.000 n=10+10)
```